+++ Minibank.Core/Domains/BankAccounts/Services/IBankAccountService.cs
@@ -8,8 +8,8 @@ namespace Minibank.Core.Domains.BankAccounts.Services
     {
         Task<BankAccount> GetById(string id, CancellationToken cancellationToken);
         Task<IEnumerable<BankAccount>> GetAll(CancellationToken cancellationToken);
-        Task<float> CalculateCommission(float amount, string fromAccountId, string toAccountId, CancellationToken cancellationToken);
-        Task TransferMoney(float amount, string fromAccountId, string toAccountId, CancellationToken cancellationToken);
+        Task<double> CalculateCommission(double amount, string fromAccountId, string toAccountId, CancellationToken cancellationToken);
+        Task TransferMoney(double amount, string fromAccountId, string toAccountId, CancellationToken cancellationToken);
         Task Create(BankAccount bankAccount, CancellationToken cancellationToken);
         Task Update(BankAccount bankAccount, CancellationToken cancellationToken);
         Task Delete(string id, CancellationToken cancellationToken);
